+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Name pluralization scheme
+    - available-bytes -> available-byte+
+    - defaults        -> default+
 - Namespace parameters under "rxtx"
 - Relicensed under MPL 2.0
 - Renamed group id from "dvlopt" to "io.helins"